扩增文心AI的功能，修复猜歌的权限问题 (#471)
